Linting and Formatting Fixes

+++ cpp-service/include/Event.h
@@ -7,19 +7,12 @@
 #include <vector>
 #include <nlohmann/json.hpp>
 
-
 /**
- * OOP PRINCIPLE: ENCAPSULATION
- * 
- * The Event class encapsulates all event data with private members.
- * External code can only access/modify data through public getters/setters,
- * allowing us to control how data is accessed and maintain data integrity.
- * 
- * Example: The 'id' cannot be changed arbitrarily - it's set through constructor
- * and protected from external modification.
+ * @file Event.h
+ * Base calendar event: core fields, attendee and notification id lists, JSON I/O, and polymorphic clone/type tag.
  */
 class Event {
-protected:  // INHERITANCE: Protected members accessible to derived classes
+protected:
     std::string id;
     std::string title;
     std::string description;
@@ -27,7 +20,7 @@ protected:  // INHERITANCE: Protected members accessible to derived classes
     std::time_t startTime;
     std::time_t endTime;
     std::string location;
-    std::string coordinatorId;  // ID of event coordinator
+    std::string coordinatorId;
     std::string fraternity;
     std::vector<std::string> attendeeIds;
     std::vector<std::string> notificationAttendeeIds;
@@ -37,14 +30,9 @@ public:
     Event();
     Event(const std::string& id, const std::string& title, const std::string& description,
           std::time_t date, const std::string& location, const std::string& coordinatorId, const std::string& fraternity);
-    
-    /**
-     * OOP PRINCIPLE: POLYMORPHISM
-     * Virtual destructor ensures proper cleanup of derived classes
-     */
+
     virtual ~Event() = default;
-    
-    // ENCAPSULATION: Getters provide read-only access
+
     std::string getId() const { return id; }
     std::string getTitle() const { return title; }
     std::string getDescription() const { return description; }
@@ -55,12 +43,10 @@ public:
     std::string getCoordinatorId() const { return coordinatorId; }
     std::string getFraternity() const { return fraternity; }
     std::vector<std::string> getAttendeeIds() const { return attendeeIds; }
-    int getAttendeeCount() const { return attendeeIds.size(); }
+    int getAttendeeCount() const { return static_cast<int>(attendeeIds.size()); }
     std::vector<std::string> getNotificationAttendeeIds() const { return notificationAttendeeIds; }
     std::vector<std::string> getNotifiedAttendeeIds() const { return notifiedAttendeeIds; }
-    
-    
-    // ENCAPSULATION: Setters provide controlled modification
+
     void setTitle(const std::string& t) { title = t; }
     void setDescription(const std::string& desc) { description = desc; }
     void setDate(std::time_t d) { date = d; }
@@ -69,31 +55,19 @@ public:
     void setLocation(const std::string& loc) { location = loc; }
     void setCoordinatorId(const std::string& cId) { coordinatorId = cId; }
     void setFraternity(const std::string& frat) { fraternity = frat; }
-    
-    // Attendee management
+
     void addAttendee(const std::string& userId);
     void removeAttendee(const std::string& userId);
     void toggleNotification(const std::string& userId, bool enabled);
     void notificationSent(const std::string& userId);
     bool isAttending(const std::string& userId) const;
-    
-    /**
-     * OOP PRINCIPLE: POLYMORPHISM
-     * Virtual methods can be overridden by derived classes to provide
-     * specialized behavior while maintaining a common interface.
-     * 
-     * Example: A RecruitmentEvent might include rush-specific fields
-     * in its JSON representation, while a PhilanthropyEvent includes
-     * fundraising goals.
-     */
+
     virtual nlohmann::json toJson() const;
     virtual void fromJson(const nlohmann::json& j);
     virtual std::string getEventType() const { return "Other"; }
-    // virtual std::string getEventDetails() const;
     virtual bool isValid() const;
-    
-    // Clone for copying polymorphic objects
+
     virtual std::shared_ptr<Event> clone() const;
 };
 
-#endif // EVENT_H
+#endif  // EVENT_H
